Save messages from message-form to database, send messages from database to js

+++ schedule/views.py
@@ -1,8 +1,10 @@
 from django.shortcuts import render
 from django.contrib.auth.models import User
 from django.contrib import messages
+from django.core.serializers import serialize
+from django.core.serializers import serialize
 import json
-from datetime import date, time
+from datetime import date, time, datetime
 
 # Form
 
@@ -21,7 +23,7 @@ class ScheduleForm(forms.ModelForm):
             "sick",
             "vacation")
 
-class MassageForm(forms.ModelForm):
+class MessageForm(forms.ModelForm):
     class Meta:
         model = Message
         fields = ("body",)
@@ -30,6 +32,7 @@ class MassageForm(forms.ModelForm):
 
 def schedule(request):
     users = list(User.objects.values('id','username', 'groups', 'is_staff'))
+    messages = list(Message.objects.values("user", "created_on" , "body"))
     schedules = list(Schedule.objects.values(
         "user",
         "date",
@@ -39,6 +42,10 @@ def schedule(request):
         "end_of_work_2",
         "sick",
         "vacation"))
+
+    for message in messages:
+        if isinstance(message["created_on"], datetime):  # Convert date to string
+            message["created_on"] = message["created_on"].strftime("%Y-%m-%d/%H:%M:%S")
     
     for schedule in schedules:
         if isinstance(schedule["date"], date):  # Convert date to string
@@ -53,46 +60,59 @@ def schedule(request):
             schedule["end_of_work_2"] = schedule["end_of_work_2"].strftime("%H:%M:%S")
 
     if request.method == "POST":
-        edit_schedule_crud = request.POST.get("edit-schedule-crud")
-        schedule_form = ScheduleForm(data=request.POST)
+        form_name = request.POST.get("form")
+        if form_name == "schedule":
+            edit_schedule_crud = request.POST.get("edit-schedule-crud")
+            schedule_form = ScheduleForm(data=request.POST)
 
-        if schedule_form.is_valid():    
-            if edit_schedule_crud == "edit":
-                print(schedule_form.cleaned_data["begin_of_work_2"])
-                if ((schedule_form.cleaned_data["begin_of_work_1"] != None
-                    and schedule_form.cleaned_data["end_of_work_1"] != None)
-                    or schedule_form.cleaned_data["sick"] == True
-                    or schedule_form.cleaned_data["vacation"] == True
-                ):
-                    existing_schedule = Schedule.objects.filter(
-                        user=schedule_form.cleaned_data["user"].id,
-                        date=schedule_form.cleaned_data["date"]
-                    ).first()
-                    print(existing_schedule)
-                    if existing_schedule:
-                        existing_schedule.begin_of_work_1 = schedule_form.cleaned_data["begin_of_work_1"]
-                        existing_schedule.end_of_work_1 = schedule_form.cleaned_data["end_of_work_1"]
-                        existing_schedule.begin_of_work_2 = schedule_form.cleaned_data["begin_of_work_2"]
-                        existing_schedule.end_of_work_2 = schedule_form.cleaned_data["end_of_work_2"]
-                        existing_schedule.sick = schedule_form.cleaned_data["sick"]
-                        existing_schedule.vacation = schedule_form.cleaned_data["vacation"]
-                        existing_schedule.save()
-                        print("shift updated")
-                           
+            if schedule_form.is_valid():    
+                if edit_schedule_crud == "edit":
+                    print(schedule_form.cleaned_data["begin_of_work_2"])
+                    if ((schedule_form.cleaned_data["begin_of_work_1"] != None
+                        and schedule_form.cleaned_data["end_of_work_1"] != None)
+                        or schedule_form.cleaned_data["sick"] == True
+                        or schedule_form.cleaned_data["vacation"] == True
+                    ):
+                        existing_schedule = Schedule.objects.filter(
+                            user=schedule_form.cleaned_data["user"].id,
+                            date=schedule_form.cleaned_data["date"]
+                        ).first()
+                        print(existing_schedule)
+                        if existing_schedule:
+                            existing_schedule.begin_of_work_1 = schedule_form.cleaned_data["begin_of_work_1"]
+                            existing_schedule.end_of_work_1 = schedule_form.cleaned_data["end_of_work_1"]
+                            existing_schedule.begin_of_work_2 = schedule_form.cleaned_data["begin_of_work_2"]
+                            existing_schedule.end_of_work_2 = schedule_form.cleaned_data["end_of_work_2"]
+                            existing_schedule.sick = schedule_form.cleaned_data["sick"]
+                            existing_schedule.vacation = schedule_form.cleaned_data["vacation"]
+                            existing_schedule.save()
+                            print("shift updated")
+                               
+                        else:
+                            schedule_form.save()
+                            print("shift created")
                     else:
-                        schedule_form.save()
-                        print("shift created")
+                        print("shift empty")
                 else:
-                    print("shift empty")
-            else:
-                Schedule.objects.filter(
-                    user = schedule_form.cleaned_data["user"].id,
-                    date = schedule_form.cleaned_data["date"]).delete()
-                print("shift deleted")
+                    Schedule.objects.filter(
+                        user = schedule_form.cleaned_data["user"].id,
+                        date = schedule_form.cleaned_data["date"]).delete()
+                    print("shift deleted")
 
+            else:
+                print("form is not valid") 
         else:
-            print("form is not valid") 
-    
+            schedule_form = ScheduleForm()
+            print("It's a new message!")
+
+            message_form = MessageForm(data=request.POST)
+
+            if message_form.is_valid():
+                print("message form is valid")
+                message_form.instance.user = request.user 
+                message_form.save()
+            else:
+                print("message form is not valid")
     else:
         schedule_form = ScheduleForm()
         print("it's not a post")
@@ -104,12 +124,7 @@ def schedule(request):
         {
             'users': json.dumps(users),
             "schedules": json.dumps(schedules),
+            "messages":json.dumps(messages),
             "schedule_form": schedule_form,
         },
     )
-
-def message(request):
-    
-
-    if request.method == "POST":
-        print("Recived a post request")
